+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog][keepachangelog], and this project adhe
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-07-21
+
 ### Added
 
 - Add option to pass a full dependency to `jspecifyVersion` (to support forks if necessary).